backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pyyaml==6.0.1
slowapi==0.1.9
